updated files structure
