+++ controllers/contacts/addContact.js
@@ -1,0 +1,9 @@
+const { ctrlWrapper } = require("../../helpers");
+const { Contact } = require("../../models");
+
+const addContact = async (req, res) => {
+  const result = await Contact.create(req.body);
+  res.status(201).json(result);
+};
+
+module.exports = ctrlWrapper(addContact);
