+++ controllers/contacts/updateContact.js
@@ -1,0 +1,15 @@
+const { HttpError, ctrlWrapper } = require("../../helpers");
+const { Contact } = require("../../models");
+
+const updateContact = async (req, res) => {
+  const { contactId } = req.params;
+  const result = await Contact.findByIdAndUpdate(contactId, req.body, {
+    new: true,
+  });
+
+  if (!result) throw HttpError(404, "Not found.");
+
+  res.json(result);
+};
+
+module.exports = ctrlWrapper(updateContact);
