+++ models/index.js
@@ -1,0 +1,5 @@
+const Contact = require("./contact");
+
+module.exports = {
+  Contact,
+};
