+++ routes/api/contacts.js
@@ -2,7 +2,7 @@ const express = require("express");
 
 const ctrl = require("../../controllers/contacts");
 const { validateBody, isValidId } = require("../../middlewares");
-const { schemas } = require("../../models/contact");
+const {addSchema, updateFavoriteSchema} = require('../../schemas')
 
 const router = express.Router();
 
@@ -10,21 +10,21 @@ router.get("/", ctrl.listContacts);
 
 router.get("/:contactId", isValidId, ctrl.getContactById);
 
-router.post("/", validateBody(schemas.addSchema), ctrl.addContact);
+router.post("/", validateBody(addSchema), ctrl.addContact);
 
 router.delete("/:contactId", isValidId, ctrl.removeContact);
 
 router.put(
   "/:contactId",
   isValidId,
-  validateBody(schemas.addSchema),
+  validateBody(addSchema),
   ctrl.updateContact
 );
 
 router.patch(
   "/:contactId/favorite",
   isValidId,
-  validateBody(schemas.updateFavoriteSchema),
+  validateBody(updateFavoriteSchema),
   ctrl.updateStatusContact
 );
