+++ schemas/index.js
@@ -1,0 +1,7 @@
+const addSchema = require("./addSchema");
+const updateFavoriteSchema = require("./updateFavoriteSchema");
+
+module.exports = {
+  addSchema,
+  updateFavoriteSchema,
+};
