+++ schemas/updateFavoriteSchema.js
@@ -1,0 +1,7 @@
+const Joi = require("joi");
+
+const updateFavoriteSchema = Joi.object({
+  favorite: Joi.boolean().required(),
+});
+
+module.exports = updateFavoriteSchema;
